Fix YNAB import travel link mapping
